Extract settings page presentation
